"debug-unloading" setting.
What is this for?  A host will keep the .clap file open as long as it has an
instance of any of its plugins open, and will unload it when the last one is
closed.  (At least, that's what Reaper does.)  During development, this lets
you verify that the .clap file was unloaded, so the next time you add a plugin,
it'll use the latest .clap you built.